Refactor: Organize repository structure
This commit reorganizes the repository into a more standard and user-friendly structure. The changes include:

-   Created `src`, `scripts`, and `docs` directories to house the Python source code, batch scripts, and documentation, respectively.
-   Moved all Python files to the `src` directory and updated their import statements to use relative imports.
-   Created an `__init__.py` file in the `src` directory to make it a Python package.
-   Moved all batch files to the `scripts` directory and updated them to call the Python scripts as modules from the `src` package.
-   Moved the `Docs.md` file to the `docs` directory.
-   Updated the `README.md` file to reflect the new file structure and provide clearer instructions for both technical and non-technical users. The README is also now in English to be more accessible.
