+++ src/image_processor.py
